Securing the docker image
This commit changes who is running the container from root to www-data which is a user already created in the base-image. This makes the image more secure as the user is not root and doesn't have root rights.

This commit also adds ./bin to the $PATH variable which makes it easyer to run the indexer scripts.