Adds more code around producing Kafka messages
Not all the messages were being delivered from this system to the Kafka queue. I'm not a 100% sure what I'm doing but I added in the code from this website https://arnaud.le-blanc.net/php-rdkafka-doc/phpdoc/rdkafka.examples-producer.html and it seems to be doing the trick

I think `$this->client->poll(0);` is the main thing, but then the deconstructor is trying 10 times to `flush` the stream. It will also throw an Exception if it can't :)